Fix route of the whole project

+++ src/Controller/UserController.php
@@ -14,7 +14,7 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/user')]
 class UserController extends AbstractController
 {
-    #[Route('/', name: 'app_user_index', methods: ['GET'])]
+    #[Route('/user', name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository, ArticleRepository $articleRepository): Response
     {
         return $this->render('user/index.html.twig', [
@@ -45,14 +45,21 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        $currentUser = $this->getUser();
+        if ($currentUser->getRoles() == ["ROLE_ADMIN, ROLE_USER"]) {
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
+    }else {
+        return $this->redirectToRoute('app_index', [], Response::HTTP_SEE_OTHER);
+    }
     }
 
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, UserRepository $userRepository): Response
     {
+        $currentUser = $this->getUser();
+        if ($currentUser->getRoles() == ["ROLE_ADMIN, ROLE_USER"]) {
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -66,34 +73,48 @@ class UserController extends AbstractController
             'user' => $user,
             'form' => $form,
         ]);
+    }else {
+        return $this->redirectToRoute('app_index', [], Response::HTTP_SEE_OTHER);
+    }
     }
 
-    #[Route('/{id}/editprofil', name: 'app_user_editprofil', methods: ['GET', 'POST'])]
+    #[Route('/editprofil/{id}', name: 'app_user_editprofil', methods: ['GET', 'POST'])]
     public function editprofil(Request $request, User $user, UserRepository $userRepository): Response
     {
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
+        $currentUser = $this->getUser();
+if ($currentUser->getId() == $user->getId()) {
+    $form = $this->createForm(UserType::class, $user);
+    $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $userRepository->save($user, true);
+    if ($form->isSubmitted() && $form->isValid()) {
+        $userRepository->save($user, true);
 
-            return $this->redirectToRoute('app_private_profil', [], Response::HTTP_SEE_OTHER);
-        }
+        return $this->redirectToRoute('app_private_profil', [], Response::HTTP_SEE_OTHER);
+    }
 
-        return $this->renderForm('user/editprofil.html.twig', [
-            'user' => $user,
-            'form' => $form,
-        ]);
+    return $this->renderForm('user/editprofil.html.twig', [
+        'user' => $user,
+        'form' => $form,
+    ]);
+}else {
+    return $this->redirectToRoute('app_index', [], Response::HTTP_SEE_OTHER);
+}
+
     }
 
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $userRepository->remove($user, true);
-        }
+        $currentUser = $this->getUser();
+if ($currentUser->getRoles() == ["ROLE_ADMIN, ROLE_USER"]) {
+    if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        $userRepository->remove($user, true);
+    }
 
-        return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
+    return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
+}else {
+            return $this->redirectToRoute('app_index', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     public function __toString(){
